Added a new resource
Hey, I have updated the grammar mistake and also added a new resource for boolean in python. I found this tutorial on boolean while looking for some resources to learn python. Topics were explained pretty well and found it to be useful and thought of adding it here. Hope this small contribution will help other learners.